+++ types/config_test.go
@@ -32,3 +32,12 @@ func TestSetBech32Prefixes(t *testing.T) {
 	require.Equal(t, sdk.GetConfig().GetBech32ConsensusAddrPrefix(), config.GetBech32ConsensusAddrPrefix())
 	require.Equal(t, sdk.GetConfig().GetBech32ConsensusPubPrefix(), config.GetBech32ConsensusPubPrefix())
 }
+
+func TestSetCoinType(t *testing.T) {
+	config := sdk.GetConfig()
+	require.Equal(t, sdk.CoinType, int(config.GetCoinType()))
+
+	SetBip44CoinType(config)
+	require.Equal(t, Bip44CoinType, int(config.GetCoinType()))
+	require.Equal(t, sdk.GetConfig().GetCoinType(), config.GetCoinType())
+}